refactor: Move InstallPickerDialog to components
